+++ src/components/Main/content/MainContent.js
@@ -1,21 +1,32 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import { API } from "../../config";
+import { API } from "../../../config";
 import styled from "styled-components";
 import ContentSection from "./ContentSection";
-import NovelFreeContent from "./content/NovelFreeContent";
 
-const SubContent = () => {
+const MainContent = () => {
   const [novel, setNovel] = useState([]);
+  const [comix, setComix] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
     setIsLoading(true);
+
     axios
       .get(API.GET_NOVELS)
       .then(response => {
         const novelData = response.data.novels;
         setNovel(novelData);
+      })
+      .catch(error => {
+        console.log("Error >>", error);
+      });
+
+    axios
+      .get(API.GET_COMICS)
+      .then(response => {
+        const comixData = response.data.comics;
+        setComix(comixData);
       })
       .catch(error => {
         console.log("Error >>", error);
@@ -30,21 +41,32 @@ const SubContent = () => {
       {isLoading ? (
         <LoadingText>Loading...</LoadingText>
       ) : (
-        <Section>
-          <TopicTitle>최근 본 작품</TopicTitle>
-          <ContentSection contentData={novel} />
-          <TopicTitle>실시간 랭킹</TopicTitle>
-          <ContentSection contentData={novel} sort={"ranking"} />
-          <NovelFreeContent />
-          <TopicTitle>지금 시리즈가 추천해요</TopicTitle>
-          <ContentSection contentData={novel} />
-        </Section>
+        <>
+          <Section>
+            <TopicTitle>실시간 랭킹</TopicTitle>
+            <ContentSection
+              topicTitle={"NOVEL"}
+              contentData={novel}
+              sort={"ranking"}
+            />
+            <ContentSection
+              topicTitle={"COMIX"}
+              contentData={comix}
+              sort={"ranking"}
+            />
+          </Section>
+          <Section>
+            <TopicTitle>지금 시리즈가 추천해요</TopicTitle>
+            <ContentSection topicTitle={"NOVEL"} contentData={novel} />
+            <ContentSection topicTitle={"COMIX"} contentData={comix} />
+          </Section>
+        </>
       )}
     </Div>
   );
 };
 
-export default SubContent;
+export default MainContent;
 
 const Div = styled.div``;
