+++ src/components/Main/content/NovelFreeContent.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { API } from "../../../config";
 import styled from "styled-components";
-import ContentSection from "../ContentSection";
+import ContentSection from "./ContentSection";
 
 const NovelFreeContent = () => {
   const [ing, seting] = useState([]);
@@ -12,10 +12,37 @@ const NovelFreeContent = () => {
 
   useEffect(() => {
     axios
+      .get(API.GET_NOVELS_FREELIST_ING)
+      .then(response => {
+        const ingData = response.data.novels;
+        seting(ingData);
+      })
+      .catch(error => {
+        console.log("Error >>", error);
+      });
+    axios
+      .get(API.GET_NOVELS_FREELIST_FREEPASS)
+      .then(response => {
+        const freepassData = response.data.novels;
+        setFreepass(freepassData);
+      })
+      .catch(error => {
+        console.log("Error >>", error);
+      });
+    axios
       .get(API.GET_NOVELS_FREELIST_FREETODAY)
       .then(response => {
         const freetodayData = response.data.novels;
         setFreetoday(freetodayData);
+      })
+      .catch(error => {
+        console.log("Error >>", error);
+      });
+    axios
+      .get(API.GET_NOVELS_FREELIST_TIMEDEAL)
+      .then(response => {
+        const timedealData = response.data.novels;
+        setTimedeal(timedealData);
       })
       .catch(error => {
         console.log("Error >>", error);
@@ -26,13 +53,22 @@ const NovelFreeContent = () => {
     <Div>
       <Section>
         <TopicTitle>무료 작품</TopicTitle>
+        <ContentSection topicTitle={"연재중"} contentData={ing} />
         <ContentSection
-          topicTitle={"연재중"}
-          contentData={freetoday}
+          topicTitle={"FREE PASS"}
+          contentData={freepass}
+          sort={"freepass"}
         />
-        <ContentSection topicTitle={"FREE PASS"} contentData={freetoday} sort={"freepass"}/>
-        <ContentSection topicTitle={"매일 10시 무료"} contentData={freetoday} sort={"freetoday"} />
-        <ContentSection topicTitle={"TIME DEAL"} contentData={freetoday} sort={"timedeal"} />
+        <ContentSection
+          topicTitle={"매일 10시 무료"}
+          contentData={freetoday}
+          sort={"freetoday"}
+        />
+        <ContentSection
+          topicTitle={"TIME DEAL"}
+          contentData={timedeal}
+          sort={"timedeal"}
+        />
       </Section>
     </Div>
   );
