--- src/components/Main/content/Ranking.js
@@ -1,32 +1,0 @@
-import React, { useState, useEffect } from "react";
-import styled from "styled-components";
-
-const Ranking = () => {
-  return (
-    <div>
-      <TopicTitle>d</TopicTitle>
-    </div>
-  );
-};
-
-export default Ranking;
-
-const TopicTitle = styled.p`
-  /* 무료 작품 */
-
-  position: absolute;
-  left: 2.04%;
-  right: 31.63%;
-  top: 0%;
-  bottom: 86.73%;
-
-  /* 토픽 */
-
-  font-family: "Apple SD Gothic Neo";
-  font-style: normal;
-  font-weight: 800;
-  font-size: 18px;
-  line-height: 22px;
-
-  color: #000000;
-`;
