+++ src/components/Main/event/EventSection.js
@@ -3,7 +3,7 @@ import axios from "axios";
 import styled from "styled-components";
 import { Carousel } from "react-responsive-carousel";
 import "react-responsive-carousel/lib/styles/carousel.min.css";
-import { API } from "../../config";
+import { API } from "../../../config";
 
 const EventSection = () => {
   const [events, setEvents] = useState([]);
