+++ src/components/Main/nav/MainNavBar.js
@@ -1,10 +1,10 @@
 import React from "react";
 import styled from "styled-components";
 import { Link, useNavigate } from "react-router-dom";
-import cookie from "../../assets/icon/Main/cookie.png";
-import logo from "../../assets/logo/logo.png";
-import profile from "../../assets/icon/Main/profile.png";
-import search from "../../assets/icon/Main/search.png";
+import cookie from "../../../assets/icon/Main/cookie.png";
+import logo from "../../../assets/logo/logo.png";
+import profile from "../../../assets/icon/Main/profile.png";
+import search from "../../../assets/icon/Main/search.png";
 
 const MainNavBar = () => {
   const navigate = useNavigate();
